Rename Resolver->Reouter; TapisCommand->TapipyCategory. Start to implment default TapipyCategory no  user-defined category exists
